修改了接口和页面效果

+++ app/src/main/java/com/simpleryo/leyotang/bean/UserInfoBean.java
@@ -305,8 +305,8 @@ public class UserInfoBean extends BaseResult {
              */
 
             private String id;
-            private int lat;
-            private int lng;
+            private double lat;
+            private double lng;
             private String provice;
             private String city;
             private String district;
@@ -323,19 +323,19 @@ public class UserInfoBean extends BaseResult {
                 this.id = id;
             }
 
-            public int getLat() {
+            public double getLat() {
                 return lat;
             }
 
-            public void setLat(int lat) {
+            public void setLat(double lat) {
                 this.lat = lat;
             }
 
-            public int getLng() {
+            public double getLng() {
                 return lng;
             }
 
-            public void setLng(int lng) {
+            public void setLng(double lng) {
                 this.lng = lng;
             }
